+++ RELEASES.md
@@ -3,7 +3,7 @@ Releases
 
  Vsn  |   Date     |   Desc
 ------|------------|----------
-[0.1.6](#0.1.6) | 2014/10/16 | [Cleanup queue ownership and status interface](#0.1.6)
+[0.1.6](#0.1.6) | 2014/10/17 | [Cleanup queue ownership and status interface](#0.1.6)
 [0.1.5](#0.1.5) | 2014/10/15 | [Serial buffering of pending requests](#0.1.5)
 [0.1.4](#0.1.4) | 2014/10/06 | [Allow enabled/disabled elysium_queue on startup](#0.1.4)
 [0.1.3](#0.1.3) | 2014/10/02 | [Add consistency to all CQL requests](#0.1.3)
@@ -13,6 +13,7 @@ Releases
 
 ### <a name="0.1.6"></a>0.1.6 Cleanup queue ownership and status interface
 
+    * Queued sessions are now {{Ip, Port}, Session_Pid} for debugging
     * Add elysium_config:is_elysium_config_enabled/1
     * Add elysium_buffer_sup to own session and pending request ets_buffers
     * Remove elysium_overload